Push process_command_send_message() up into a parent class. Also cleanup some leftover cruft.
