+++ go/apps/dialogue/vumi_app.py
@@ -109,35 +109,3 @@ class DialogueApplication(JsBoxApplication):
                 to_addr = contact.addr_for(delivery_class)
                 yield self.send_first_dialogue_message(
                     to_addr, contact, conv)
-
-    @inlineCallbacks
-    def process_command_send_message(self, user_account_key, conversation_key,
-                                     **kwargs):
-        command_data = kwargs.pop('command_data')
-        if kwargs:
-            log.info("Received unexpected command args: %s" % (kwargs,))
-        conv = yield self.get_conversation(user_account_key, conversation_key)
-        if conv is None:
-            log.warning("Cannot find conversation '%s' for user '%s'." % (
-                conversation_key, user_account_key))
-            return
-
-        log.info('Processing send_message: %s' % kwargs)
-        to_addr = command_data['to_addr']
-        content = command_data['content']
-        msg_options = command_data['msg_options']
-        in_reply_to = msg_options.pop('in_reply_to', None)
-        self.add_conv_to_msg_options(conv, msg_options)
-        if in_reply_to:
-            msg = yield self.vumi_api.mdb.get_inbound_message(in_reply_to)
-            if msg:
-                # We can't override transport_name in reply_to(), so we set it
-                # on the message we're replying to.
-                msg['transport_name'] = msg_options['transport_name']
-                yield self.reply_to(msg, content)
-            else:
-                log.warning('Unable to reply, message %s does not exist.' % (
-                    in_reply_to))
-        else:
-            yield self.send_to(
-                to_addr, content, endpoint='default', **msg_options)
