+++ go/apps/http_api/vumi_app.py
@@ -123,35 +123,6 @@ class StreamingHTTPWorker(GoApplicationWorker):
         return conversation.config.get('http_api', {}).get(key)
 
     @inlineCallbacks
-    def process_command_send_message(self, user_account_key, conversation_key,
-                                     **kwargs):
-        conv = yield self.get_conversation(user_account_key, conversation_key)
-        if conv is None:
-            log.warning("Cannot find conversation '%s' for user '%s'." % (
-                conversation_key, user_account_key))
-            return
-
-        command_data = kwargs['command_data']
-        log.info('Processing send_message: %s' % kwargs)
-        to_addr = command_data['to_addr']
-        content = command_data['content']
-        msg_options = command_data['msg_options']
-        self.add_conv_to_msg_options(conv, msg_options)
-        in_reply_to = msg_options.pop('in_reply_to', None)
-        if in_reply_to:
-            msg = yield self.vumi_api.mdb.get_inbound_message(in_reply_to)
-            if msg:
-                # We can't override transport_name in reply_to(), so we set it
-                # on the message we're replying to.
-                msg['transport_name'] = msg_options['transport_name']
-                yield self.reply_to(msg, content)
-            else:
-                log.warning('Unable to reply, message %s does not exist.' % (
-                    in_reply_to))
-        else:
-            yield self.send_to(to_addr, content, **msg_options)
-
-    @inlineCallbacks
     def consume_user_message(self, message):
         msg_mdh = self.get_metadata_helper(message)
         user_api = msg_mdh.get_user_api()
