+++ go/apps/surveys/tests/test_vumi_app.py
@@ -326,34 +326,6 @@ class TestSurveyApplication(AppWorkerTestCase):
         self.assertEqual(sent_msg['in_reply_to'], msg['message_id'])
 
     @inlineCallbacks
-    def test_process_command_send_message_in_reply_to_bad_transport_name(self):
-        yield self.start_conversation(self.conversation)
-        batch_id = yield self.conversation.get_latest_batch_key()
-        msg = self.mkmsg_in(message_id=uuid.uuid4().hex, transport_name="bad")
-        yield self.store_inbound_msg(msg)
-        command = VumiApiCommand.command(
-            'worker', 'send_message',
-            user_account_key=self.user_account.key,
-            conversation_key=self.conversation.key,
-            command_data={
-                u'batch_id': batch_id,
-                u'content': u'foo',
-                u'to_addr': u'to_addr',
-                u'msg_options': {
-                    u'transport_name': u'smpp_transport',
-                    u'in_reply_to': msg['message_id'],
-                    u'transport_type': u'sms',
-                    u'from_addr': u'default10080',
-                }
-            })
-        yield self.app.consume_control_command(command)
-        [sent_msg] = self.get_dispatched_messages()
-        self.assertEqual(sent_msg['to_addr'], msg['from_addr'])
-        self.assertEqual(sent_msg['content'], 'foo')
-        self.assertEqual(sent_msg['in_reply_to'], msg['message_id'])
-        self.assertEqual(sent_msg['transport_name'], 'smpp_transport')
-
-    @inlineCallbacks
     def test_closing_menu_if_unavailable(self):
         poll_id = 'poll-%s' % (self.conversation.key,)
         config = yield self.pm.get_config(poll_id)
